requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
coverage==7.11.3
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
fonttools==4.60.1
googlemaps==4.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.0.5
langchain-core==1.0.4
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.42
markdown-it-py==4.0.0
matplotlib==3.10.7
mdurl==0.1.2
mypy_extensions==1.1.0
numpy==2.3.4
openai==2.8.0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.15.4
pyparsing==3.2.5
pytest==9.0.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tabulate==0.9.0
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
xxhash==3.6.0
zstandard==0.25.0
